back/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0
python-dotenv==1.0.0
redis==4.6.0
flask-caching==2.1.0 